Resolve "Add Authorization Check to Delete Ad View"

Closes #24 and #18

See merge request !20

+++ webroot/textbook_app/views.py
@@ -81,6 +81,9 @@ def profile(request):
     return render(request, 'textbook_app/profile.html', context)
 
 def ad_delete(request, ad_id):
+    ad = get_object_or_404(Ad, pk=ad_id)
+    if ad.poster.id != request.user.id:
+        return HttpResponseRedirect(reverse('textbook_app:profile'))
     ad_tobe_deleted = get_object_or_404(Ad, pk=ad_id)
     ad_tobe_deleted.delete()
     return HttpResponseRedirect(reverse('textbook_app:profile'))
@@ -91,6 +94,8 @@ def ad_delete(request, ad_id):
 def ad_new_or_edit(request, isEditAd, ad_id):
     if isEditAd:
         ad = get_object_or_404(Ad, pk=ad_id)
+        if ad.poster.id != request.user.id:
+            return HttpResponseRedirect(reverse('textbook_app:profile'))
     if request.method == 'POST':
         textbookForm = TextbookForm(request.POST)
         if isEditAd:
